Move ledger-api-bench-tool classes to individual files
CHANGELOG_BEGIN
CHANGELOG_END
