+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/metrics/ServiceLevelObjective.scala
@@ -1,0 +1,9 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.api.benchtool.metrics
+
+trait ServiceLevelObjective[MetricValueType <: MetricValue] {
+  def isViolatedBy(metricValue: MetricValueType): Boolean
+  def formatted: String
+}
